Changed ints to longs

+++ src/main/java/software/aoc/day02/a/model/Range.java
@@ -1,4 +1,4 @@
 package software.aoc.day02.a.model;
 
-public record Range(int from, int to) {
+public record Range(long from, long to) {
 }
